set up s3 R2 cloudflare

+++ src/services/StorageEngine.ts
@@ -2,7 +2,9 @@ import fs, {RmOptions} from "fs";
 import path from "path";
 
 function remove(path: string, options?: RmOptions) {
-	return fs.rmSync(path, options);
+	if (isExist(path)) {
+		return fs.rmSync(path, options);
+	}
 }
 
 function joinPath(...paths: string[]): string {
